* 'develop' of github.com:MetaMask/metamask-mobile: (51 commits)
  Feature/confusables (#2464)
  fix typeface on login text field (#2610)
  Replace controller context (#2416)
  Fix adding custom token in custom network (#2590)
  only add custom tokens if not in mainnet (#2470)
  Address yarn lints (#2524)
  Upgrade .nvmrc to node v14 (#2588)
  Swaps: Add cache thresholds configuration (#2514)
  Swaps: BSC Support (#2468)
  Use node 14 (#2539)
  resolve isENS without case sensitivity (#2545) (#2568)
  Revert "resolve isENS without case sensitivity (#2545)" (#2566)
  resolve isENS without case sensitivity (#2545)
  Bump versioncode (#2558)
  v2.2.0 (#2555)
  Include decimalsToShow in balanceToFiatNumber (#2547)
  Bug fix/sync import time (#2554)
  bundle update (#2549)
  Fix analytics try catch (#2546)
  Only get nonce from the network if the feature is enabled (#2543)
  ...

+++ app/components/Views/Login/index.js
@@ -85,7 +85,8 @@ const styles = StyleSheet.create({
 		flexDirection: 'column'
 	},
 	label: {
-		fontSize: 14,
+		color: colors.black,
+		fontSize: 16,
 		marginBottom: 12,
 		...fontStyles.normal
 	},
@@ -114,10 +115,16 @@ const styles = StyleSheet.create({
 	biometryLabel: {
 		flex: 1,
 		fontSize: 16,
+		color: colors.black,
 		...fontStyles.normal
 	},
 	biometrySwitch: {
 		flex: 0
+	},
+	input: {
+		...fontStyles.normal,
+		fontSize: 16,
+		paddingTop: 2
 	},
 	cant: {
 		width: 280,
@@ -278,7 +285,8 @@ class Login extends PureComponent {
 			// Restore vault with user entered password
 			await KeyringController.submitPassword(this.state.password);
 			const encryptionLib = await AsyncStorage.getItem(ENCRYPTION_LIB);
-			if (encryptionLib !== ORIGINAL) {
+			const existingUser = await AsyncStorage.getItem(EXISTING_USER);
+			if (encryptionLib !== ORIGINAL && existingUser) {
 				await recreateVaultWithSamePassword(this.state.password, this.props.selectedAddress);
 				await AsyncStorage.setItem(ENCRYPTION_LIB, ORIGINAL);
 			}
@@ -477,6 +485,7 @@ class Login extends PureComponent {
 					<View style={styles.areYouSure}>
 						<Text style={[styles.heading, styles.delete]}>{strings('login.type_delete')}</Text>
 						<OutlinedTextField
+							style={styles.input}
 							autoFocus
 							returnKeyType={'done'}
 							onChangeText={this.checkDelete}
@@ -511,6 +520,7 @@ class Login extends PureComponent {
 						<View style={styles.field}>
 							<Text style={styles.label}>{strings('login.password')}</Text>
 							<OutlinedTextField
+								style={styles.input}
 								placeholder={'Password'}
 								testID={'login-password-input'}
 								returnKeyType={'done'}
